Исправила на Integer поле createdAt

+++ src/main/java/ru/skypro/homework/dto/CommentDto.java
@@ -7,7 +7,7 @@ public class CommentDto {
     private Integer author;
     private String authorImage;
     private String authorFirstName;
-    private Long createdAt;
+    private Integer createdAt;
     private Integer pk;
     private String text;
 }
